+++ src/main/java/com/example/lastproject/common/dto/AuthUser.java
@@ -1,4 +1,4 @@
-package com.example.lastproject.domain.auth.entity;
+package com.example.lastproject.common.dto;
 
 import com.example.lastproject.domain.user.enums.UserRole;
 import lombok.Builder;
@@ -12,6 +12,11 @@ import java.util.List;
 @Getter
 @Builder
 public class AuthUser {
+
+    /*
+    AuthUser 는 인증, 인가된 후의 유저임을 알 수 있는 객체 -> 엔티티가 아닌 DTO
+    엔티티로 생성됐을 때 JPA 에서 관리해야 할 필요성이 생기게 됨
+     */
 
     private final Long userId;
     private final String email;
