+++ src/main/java/com/example/lastproject/common/exception/GlobalExceptionHandler.java
@@ -1,8 +1,7 @@
-package com.example.lastproject.common;
+package com.example.lastproject.common.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.http.converter.HttpMessageNotReadableException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
