+++ src/main/java/com/example/lastproject/config/BatchConfig.java
@@ -1,6 +1,6 @@
 package com.example.lastproject.config;
 
-import com.example.lastproject.common.CustomException;
+import com.example.lastproject.common.exception.CustomException;
 import com.example.lastproject.common.enums.ErrorCode;
 import com.example.lastproject.domain.item.entity.Item;
 import com.example.lastproject.domain.item.repository.ItemRepository;
