+++ src/main/java/com/example/lastproject/config/JwtSecurityFilter.java
@@ -1,6 +1,6 @@
 package com.example.lastproject.config;
 
-import com.example.lastproject.domain.auth.entity.AuthUser;
+import com.example.lastproject.common.dto.AuthUser;
 import com.example.lastproject.domain.user.enums.UserRole;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.ExpiredJwtException;
