+++ src/main/java/com/example/lastproject/domain/chat/service/ChatRoomServiceImpl.java
@@ -1,9 +1,9 @@
 package com.example.lastproject.domain.chat.service;
 
-import com.example.lastproject.common.CustomException;
+import com.example.lastproject.common.exception.CustomException;
 import com.example.lastproject.common.annotation.LogisticsNotify;
 import com.example.lastproject.common.enums.ErrorCode;
-import com.example.lastproject.domain.auth.entity.AuthUser;
+import com.example.lastproject.common.dto.AuthUser;
 import com.example.lastproject.domain.chat.dto.ChatRoomResponse;
 import com.example.lastproject.domain.chat.entity.ChatRoom;
 import com.example.lastproject.domain.chat.repository.ChatRoomRepository;
