+++ src/main/java/com/example/lastproject/domain/item/service/ItemServiceImpl.java
@@ -1,6 +1,6 @@
 package com.example.lastproject.domain.item.service;
 
-import com.example.lastproject.common.CustomException;
+import com.example.lastproject.common.exception.CustomException;
 import com.example.lastproject.common.enums.ErrorCode;
 import com.example.lastproject.domain.item.dto.response.ItemResponse;
 import com.example.lastproject.domain.item.entity.Item;
