+++ src/main/java/com/example/lastproject/domain/likeitem/service/LikeItemServiceImpl.java
@@ -1,12 +1,10 @@
 package com.example.lastproject.domain.likeitem.service;
 
-import com.example.lastproject.common.CustomException;
+import com.example.lastproject.common.exception.CustomException;
 import com.example.lastproject.common.enums.ErrorCode;
-import com.example.lastproject.domain.auth.entity.AuthUser;
+import com.example.lastproject.common.dto.AuthUser;
 import com.example.lastproject.domain.item.entity.Item;
-import com.example.lastproject.domain.item.repository.ItemRepository;
 import com.example.lastproject.domain.item.service.ItemService;
-import com.example.lastproject.domain.item.service.ItemServiceImpl;
 import com.example.lastproject.domain.likeitem.dto.response.LikeItemResponse;
 import com.example.lastproject.domain.likeitem.entity.LikeItem;
 import com.example.lastproject.domain.likeitem.repository.LikeItemRepository;
