+++ src/main/java/com/example/lastproject/domain/notification/controller/NotificationController.java
@@ -1,6 +1,6 @@
 package com.example.lastproject.domain.notification.controller;
 
-import com.example.lastproject.domain.auth.entity.AuthUser;
+import com.example.lastproject.common.dto.AuthUser;
 import com.example.lastproject.domain.notification.dto.response.NotificationListResponse;
 import com.example.lastproject.domain.notification.service.NotificationService;
 import lombok.RequiredArgsConstructor;
